Added most recent files
v0.3 release

Stable with optimistations and new features

+++ autopost.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import time
+import instaReposter
 from os import listdir
 from os.path import isfile, join
 from random import randint
@@ -8,6 +9,7 @@ from InstagramAPI import InstagramAPI
 
 d = os.getcwd()
 ImagePath = d + "\\Images"
+hashtags = "#meme #memes #dankmemes #funny #funnymemes #memesdaily #lol #edgymemes #dank #offensivememes #lmao #follow #like #bhfyp #humor #dankmeme #reddit #edgy #comedy #fun #instagram #cringe #offensive #sad #memestagram"
 
 def main():  
         os.chdir(ImagePath)
@@ -18,6 +20,7 @@ def main():
 
 
 def instaLogin():
+
         os.chdir("..")
         f = open("instaLogin.txt", "r")
         igUser = f.readline()
@@ -25,40 +28,48 @@ def instaLogin():
         f.close()
         igapi = InstagramAPI(igUser, igPassword)
         igapi.login()  # login
-        print("Logged in")
+        print("Logged into Instagram successfully")
         os.chdir(ImagePath)
         return igapi
 
-def postIt():
-        igapi = instaLogin()
-        os.chdir(ImagePath)
-        ListFiles = sorted([f for f in listdir(ImagePath) if isfile(join(ImagePath, f))])
-        print(ListFiles)
-        print("Total number of photos in this folder:" + str(len(ListFiles)))  
-        
-        
-        for i, _ in enumerate(ListFiles):
-                photo = ListFiles[i]
-                print("----------------------------------------------------------")
-                print("Progress: " + str([i + 1]) + " of " + str(len(ListFiles)))
-                credit = os.path.splitext(photo)[0]
-                IGCaption = ("Originally posted by: /u/" + credit + " on Reddit" ".\n.\n.\n.\n.\n.\n.\n.\nTags:\n\n#meme #memes #dankmemes #funny #funnymemes #memesdaily #lol #edgymemes #dank #offensivememes #lmao #follow #like #bhfyp #humor #dankmeme #reddit #edgy #comedy #fun #instagram #cringe #offensive #sad #memestagram")
-                print("Now uploading this photo to Instagram: " + photo)
-                igapi.uploadPhoto(photo, caption=IGCaption, upload_id=None)
-                os.remove(photo)
-                if len(listdir(ImagePath)) == 0:
-                        print("Posting Halted")
-                        time.sleep(5)
-                        print("Posting restarted")
-                        main()
-                else:
-                        timer()
-                
 
+
+def postIt():
+
+        igapi = instaLogin()
+        
+        try:
+                while True:
+                        os.chdir(ImagePath)
+                        ListFiles = sorted([f for f in listdir(ImagePath) if isfile(join(ImagePath, f))])
+                        print(ListFiles)
+                        print("Total number of photos in this folder:" + str(len(ListFiles)))  
+        
+                        if len(listdir(ImagePath)) != 0:
+                                for i, _ in enumerate(ListFiles):
+                                        photo = ListFiles[i]
+                                        print("----------------------------------------------------------")
+                                        print("Progress: " + str([i + 1]) + " of " + str(len(ListFiles)))
+                                        credit = os.path.splitext(photo)[0]
+                                        IGCaption = ("Originally posted by: /u/" + credit + " on Reddit" ".\n.\n.\n.\n.\n.\n.\n.\nTags:\n\n" + hashtags)
+                                        print("Now uploading this photo to Instagram: " + photo)
+                                        igapi.uploadPhoto(photo, caption=IGCaption, upload_id=None)
+                                        os.remove(photo)
+                                        timer()
+                        elif len(listdir(ImagePath)) == 0:
+                                #image grab
+                                instaReposter.multiProcessor(2)
+                                time.sleep(10)
+        except KeyboardInterrupt:
+                print("Program interrupted")
+                pass          
+
+        
 
 def timer():
     n = randint(1400, 1800)
     print("----------------------------------------------------------")
+    
     for remaining in range(n, 0, -1):
         sys.stdout.write("\r")
         sys.stdout.write("{:2d} seconds remaining to next upload.".format(remaining))
